Refactor: organize DAO tests
