+++ test/dao/games/games_mem_test.dart
@@ -1,6 +1,6 @@
 import 'package:ayso_ks/dao/mem/games.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 import 'games_spec.dart';
 
 void main() {
