+++ test/dao/games/games_spec.dart
@@ -1,7 +1,7 @@
 import 'package:ayso_ks/dao/games.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 
 typedef GamesDaoGenerator = GamesDAO Function();
