+++ test/dao/settings/settings_mem_test.dart
@@ -1,6 +1,6 @@
 import 'package:ayso_ks/dao/mem/settings.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 import 'settings_spec.dart';
 
 void main() {
