+++ test/dao/settings/settings_spec.dart
@@ -5,7 +5,7 @@ import 'package:ayso_ks/dao/teams.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 
 typedef SettingsDaoGenerator = SettingsDAO Function();
