+++ test/dao/teams/teams_mem_test.dart
@@ -1,6 +1,6 @@
 import 'package:ayso_ks/dao/mem/teams.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 import 'teams_spec.dart';
 
 void main() {
