+++ test/dao/week_cache/week_cache_mem_test.dart
@@ -1,6 +1,6 @@
 import 'package:ayso_ks/dao/mem/week_cache.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 import 'week_cache_spec.dart';
 
 void main() {
