+++ test/dao/week_cache/week_cache_spec.dart
@@ -1,7 +1,7 @@
 import 'package:ayso_ks/dao/week_cache.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../mock/test_data.dart';
+import '../../mock/test_data.dart';
 
 typedef WeekCacheDaoGenerator = WeekCacheDAO Function();
